Refactor style for information section
Due to sizing style of previous html to contain images in their parent
div the paragraph section was overflowing into the quote section.  By
adding a container around just the picture and refactoring the css the
overflow has been fixed and the picture is still restricted to its
parent container.